fix: fix how messages are used in web-sdk
this is pretty big but it makes sense to do it sooner than later
we've mixed up how we use messages (messages sent to us by dreams and messages sent
to dreams by the partner) and it was a mess to figure out easily what's
happening.
Hence this change.
That's just first commit, a cleanup will follow